+++ CHANGELOG.md
@@ -8,6 +8,9 @@
   - Fix display of buttons on the search bar
   - Highlight charged balance amounts that are different from their final balance
   - Display negative balances on the Current Balance page in red
+- **Stastics Changes**
+  - Add a new statistics page that displays a 52-week graph of a set of
+    searched entries
 - **Gem Changes**
   - Update to ruby 3.1.2
   - Update to rails 6.1.4.4